relocate business logic from views to services
